requirements.txt
astroid==3.3.5
cfgv==3.4.0
dill==0.3.9
distlib==0.3.9
filelock==3.16.1
identify==2.6.2
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pstest==0.1.5
psycopg2==2.9.10
pylint==3.3.1
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.36
tomlkit==0.13.2
typing_extensions==4.12.2
virtualenv==20.27.1
